CRISPR_AI_Deployment/requirements.txt
streamlit
biopython
pandas
matplotlib
reportlab
scikit-learn
requests
